extracted sealed module into separate file
